+++ packages/sui-lint/CHANGELOG.md
@@ -1,5 +1,23 @@
 # CHANGELOG
 
+# 4.22.0 (2022-06-21)
+
+
+### Features
+
+* **packages/sui-lint:** Upgrade dependencies for sui-lint with bugfixes ([ac40620](https://github.com/SUI-Components/sui/commit/ac40620f5ada96f7338df442fd65cd320af910b8))
+
+
+
+# 4.21.0 (2022-06-14)
+
+
+### Features
+
+* **packages/sui-lint:** Upgrade dependencies ([6ecbb27](https://github.com/SUI-Components/sui/commit/6ecbb27a3cc17c2c99b34d67f4355c25d10d9d21))
+
+
+
 # 4.20.0 (2022-06-10)